Aggiunti repositories e aggiornate entità con nuovi attributi: - Creati repository per tutte le entità (Utente, Videogioco, Carrello, Ordine, DettaglioCarrello, DettaglioOrdine) - Aggiunto campo 'attivo' in Carrello con relativo uniqueConstraint - Aggiunto campo 'nascosto' in Videogioco - Modificato 'prezzoUnitario' in DettaglioCarrello da double a float

+++ src/main/java/com/example/gamestore/entities/DettaglioCarrello.java
@@ -40,7 +40,7 @@ public class DettaglioCarrello {
 
     @Basic
     @Column(name = "prezzo_unitario")
-    private double prezzoUnitario;
+    private float prezzoUnitario;
 
 
 }
